Upgrade script for Windows
This script will download a release or snapshot build, extract it
and perform an in-place upgrade.

Known limitations:
It will not take care of addons (.jar or .kar files for offline
installation) in the addons folder.  It will also not be able to update
itself before performing the actual update, or check the current openHAB
version.

Signed-off-by: Yannick Schaus <habpanel@schaus.net>