don't use same feature repo content twice
If the addon KAR file is used there exist two different feature repo
that contains the same content. So every feature exist twice.
This is caused because of different GAV for the same XML.
The one feature repo is created by the feature packaging and the other
one is created by the KAR packaging.
The KAR packaging already included also the feature packaging, so it
should be done only once.

As we must not break existing installations we need to keep the GAV of
the online addons.

Fixes: https://github.com/openhab/openhab-distro/issues/519
Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>