Revert to java 8 to not alienate all binding developers (#1063)
Since there is only 1 version of the eclipse setup making it install java 11 would mean it's completely useless for 2.5.x binding developers. So this changes reverts it. Ideally it should be an option the user can choose from. However I have no idea how to do that with the black magic of oomph :disappointed: 

Signed-off-by: Hilbrand Bouwkamp <hilbrand@h72.nl>