Use XML instead of properties file for log4j2 configuration
Note that Karaf processes the log4j2 configuration to add the sshd logger and this processing of XML elements is case sensitive.
It will throw exceptions if lower case element names are used or when the root logger is at the end of the list of loggers.

When new loggers are added, the XML configuration is serialized such that:
* The <Configuration> element is added to the same line as the <?xml> element
* Attributes are sorted so therefor the 'name' is at the end of the appender/logger lines

Related to #516

Signed-off-by: Wouter Born <github@maindrain.net>