Use JSON Storage in Demo App
It simplifies debugging because it is easier to check what is stored by checking the .json files.
Because we also use it in the distro it allows for easier reproducing issues or reusing existing .json storage files.
If there are breaking changes in the stored data these will also be easier to patch manually.
It is probably also not easy to patch OH2 mapdb storage files when contributors start developing with OH3.

Signed-off-by: Wouter Born <github@maindrain.net>