Remove obsolete addons from managed configuration (#1530)
It has been reported that the removal of the javascript transformation in openHAB 4 results in warnings because the add-on could not be found during installation. Since the add-on is no longer present, it does not show up in the UI and can't be uninstalled from there. This adds an update instruction to remove it from the addons.config.

Signed-off-by: Jan N. Klug <github@klug.nrw>