Create Maven "dev" profile
The "dev" profile modifies the Maven build so that only the "openhab" module is built, and it builds to a folder instead of an archive. Each subsequent build will delete the "runtime" folder inside the built folder, so that subsequent builds can be done without "clean", allowing the configuration of the "OH installation" to persist across rebuilds. It can't be used with "mvn install", use "mvn package -Pdev" instead. The built folder is a ready to start OH installation, use "start_debug" to if you want to connect a debugger.

Signed-off-by: Ravi Nadahar <nadahar@rediffmail.com>